Implement ResponseStop()
